requirements.txt
flask_login
flask_wtf
flask_bootstrip
flask_caching
simplejson
